refactor: extract versionMappings to a versions.js file, where we'll track more version things
